[vcr] Account for known_peers when making requests
Previously, if hostPort wasn't specified, the VCR proxy service would end up
making a request back to the peer which made the request and that would fail
because that peer doesn't know how to respond to that request. The fix
involves tracking the known peers of a TChannel at the time the request was
made and sending it to the VCR proxy service.

Actually resolves #138.