Fix unconsumed StreamClosedError logs
The tests were logging unconsumed StreamClosedErrors a lot. The IOLoop was
getting stopped before it could tell the future to call the callback, so the
future exception was never getting consumed.

I think this will solve the same problem for uses of `IOLoop.run_sync` that
time out as well.