peer selection: Never choose ephemeral peers (#449)
This changes peer selection to never choose ephemeral peers to send requests,
even if that is the only remaining peer.

Additionally, it loosens up the restrictions for what counts as an ephemeral
peer. Either the host can be `0.0.0.0` or the port can be `0` for a peer to be
an ephemeral. This is because Python TChannels that are not listening
advertise themselves as `127.0.0.1:0` (🙁).