Allow peerGroups to maintain one-off hostports without adding them to heap selection
Summary: This PR modifies the peerGroup logic so that if a request is
sent directly to a Peer (through a hostport passed through the call
stack) instead of adding that peer to the peer heap, we will create it,
and only keep track of it in the `peers` map (not the peer heap).  This
means we will still maintain the peer, but we won't select it for the
regular hyperbahn instance.

In addition to this, you can 'upgrade' one of these peers to be added to
the heap if you explicitly call `add` with the hostport of that peer

Task: #464