Run _close_cb if connection is already closed
Summary: Some connections are closing but not getting removed from
the peer list of connections, this is possible because of a race
condition between registering a close function and the connection being
disconnected.  This diff makes it so if you register the close callback
and the connection is already closed, it will immediately run the
callback.