peer: Don't send requests to incoming peers
This alters peer selection to stop considering peers that send incoming
requests as valid candidates to send outgoing requests to. This was
previously partially fixed in #449 where we stopped considering
ephemeral peers, but non-ephemeral peers are still an issue.

The only peers we should consider for outgoing requests are those we
explicitly added to the peer heap.

Resolves #478