Ignore incoming peers only if they're unknown (#482)
The fix for #478 in f8fa6dc was broken. It ignored any peer that was
providing incoming connections only. This doesn't work if the peer
connecting to you is a known router being advertised to.

The correct fix, as suggested by @prashantv in the original issue is to
ignore incoming peers from the peer heap if they're not already known.
This reverts the change made to the peer selection logic in f8fa6dc
while keeping the test and adds the suggested fix instead. This also
adds a test that fails both, without this change and without f8fa6dc.